Move load functionality  to Storage.java
