+++ src/main/java/DiligentPenguin.java
@@ -73,61 +73,6 @@ public class DiligentPenguin {
         }
     }
 
-    private static Task readTask(String line) throws ChatBotException, DateTimeParseException {
-        Task task;
-        String[] parts = line.split("\\|");
-        String type = parts[0].trim();
-        boolean isDone = parts[1].trim().equals("X");
-        String description = parts[2].trim();
-        String deadline = (parts.length == 4) ? parts[3].trim() : "";
-        String startTime = (parts.length == 5) ? parts[3].trim() : "";
-        String endTime = (parts.length == 5) ? parts[4].trim() : "";
-        // The above code is inspired by a conversation with chatGPT
-        switch (type) {
-        case "T":
-            task = new ToDo(description);
-            if (isDone) {
-                task.setDone();
-            }
-            return task;
-        case "D":
-            LocalDate formattedDeadline = LocalDate.parse(deadline, Task.inputFormatter);
-            task = new Deadline(description, formattedDeadline);
-            if (isDone) {
-                task.setDone();
-            }
-            return task;
-        case "E":
-            LocalDate formattedStartTime = LocalDate.parse(startTime, Task.inputFormatter);
-            LocalDate formattedEndTime = LocalDate.parse(endTime, Task.inputFormatter);
-            task = new Event(description, formattedStartTime, formattedEndTime);
-            if (isDone) {
-                task.setDone();
-            }
-            return task;
-        default:
-            throw new ChatBotException("Oops! It seems that the format of the saved file is wrong/corrupted!");
-        }
-    }
-
-    public void loadTaskList() throws ChatBotException, FileNotFoundException {
-        File file = new File(tasksFilePath);
-        Scanner scanner = new Scanner(file);
-        String taskDescription = "";
-        while (scanner.hasNext()) {
-            try {
-                taskDescription = scanner.nextLine();
-                tasks.add(DiligentPenguin.readTask(taskDescription));
-            } catch (DateTimeParseException e) {
-                System.out.println("Oops! There's something wrong with the datetime format of this line: ");
-                System.out.println(taskDescription);
-            }
-        }
-        if (tasks.isEmpty()) {
-            throw new ChatBotException("No data found!");
-        }
-    }
-
     public void mark(int i) throws ChatBotException {
         try {
             tasks.finish(i);
@@ -165,7 +110,7 @@ public class DiligentPenguin {
         Scanner scanner = new Scanner(System.in);
         this.greet();
         try {
-            this.loadTaskList();
+            this.storage.loadTaskList(tasks);
             this.storage.save(tasks);
             System.out.println("I have successfully loaded previous task list for you!");
             this.list();
